feat: extract credential schema from blocks
